pyproject.toml
[build-system]
requires = ["setuptools>=74.0.0", "wheel>=0.44.0"]
build-backend = "setuptools.build_meta"

[project]
name = "MarsBench"
version = "0.1.0"
description = "A benchmarking suite for composition in machine learning models"
readme = "README.md"
requires-python = ">=3.6"
license = {text = "MIT"}
authors = [
    {name = "MarsBench Authors"}
]
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "pytorch-lightning>=2.4.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.1.0",
    "pandas>=2.2.2",
    "pillow>=10.2.0",
    "segmentation-models-pytorch>=0.4.0",
    "wandb>=0.18.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "black>=24.8.0",
    "flake8>=4.0.1",
    "isort>=5.10.0",
    "pre-commit>=3.8.0",
    "flake8-pyproject",
]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
filterwarnings = [
    "ignore::ResourceWarning",
    "ignore::DeprecationWarning",
    "ignore:The `srun` command is available on your system but is not used:UserWarning",
    "ignore:.*cuda initialization.*:UserWarning",
    "ignore:.*overflow encountered in exp.*:RuntimeWarning",
]

[tool.black]
line-length = 120
target-version = ["py36", "py37", "py38", "py39"]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | legacy_code
)/
'''

[tool.isort]
profile = "black"
force_single_line = true
skip_glob = ["legacy_code/**"]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503", "E402", "F403"]
ignore-long-lines = "^\\s*#|^\\s*\"{3}|^\\s*'{3}"
per-file-ignores = "tests/*.py:F401"
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "legacy_code",
]
